% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{f_ob}
\alias{f_ob}
\title{Objective function of the optimization problem}
\usage{
f_ob(x, C_bar, S_vec)
}
\arguments{
\item{x}{a vector of size \eqn{m} with current solution}

\item{C_bar}{a matrix of size \eqn{nT x m}}

\item{S_vec}{a vector of size \eqn{nT}}
}
\value{
the value of the objective function at the current solution
}
\description{
Objective function of the optimization problem
}
