% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_funny_matrix}
\alias{get_funny_matrix}
\title{Funny matrices}
\usage{
get_funny_matrix(k)
}
\arguments{
\item{k}{number of columns}
}
\value{
a \eqn{(k+1) x k} matrix
}
\description{
Building blocks for CS matrices
}
