% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver_funcs.R
\name{grad_f}
\alias{grad_f}
\title{Gradient of the objective function}
\usage{
grad_f(x, C_bar, S_vec)
}
\arguments{
\item{x}{a vector of size \eqn{m} with current solution}

\item{C_bar}{a matrix of size \eqn{nT x m}}

\item{S_vec}{a vector of size \eqn{nT}}
}
\value{
the gradient of the objective function at the current solution
}
\description{
Gradient of the objective function
}
