% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sym_mat_2_vec}
\alias{sym_mat_2_vec}
\title{Symmetric matrix to vector}
\usage{
sym_mat_2_vec(M)
}
\arguments{
\item{M}{symmetric matrix}
}
\value{
a vector
}
\description{
Returns the lower triangle (without the diagonal) of a symmetric matrix as a vector
}
