% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{vec_2_sym_mat}
\alias{vec_2_sym_mat}
\title{Vector to symmetric matrix}
\usage{
vec_2_sym_mat(v)
}
\arguments{
\item{V}{vector}
}
\value{
an \eqn{n x n} symmetric matrix with diagonal \code{rep(0, n)}
}
\description{
Vector to symmetric matrix
}
